SIParser cleanup. Cleans up SIParser::LoadPrivateTypes()
git-svn-id: http://svn.mythtv.org/svn/trunk@9096 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/siparser.h
@@ -146,8 +146,7 @@ class SIParser : public QObject
     // Fixes for various DVB Network Spec Deviations
     void LoadDVBSpecDeviations(uint16_t NetworkID);
 
-    void LoadPrivateTypes(uint16_t networkID);
-
+    void LoadPrivateTypes(uint networkID);
 
     // MPEG Transport Parsers (ATSC and DVB)
     tablehead_t       ParseTableHead(uint8_t* buffer, int size);
